# renren-admin/src/main/java/io/renren/modules/sys/controller/SysUserController.java
/**
 * Copyright 2018 人人开源 http://www.renren.io
 * <p>
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy of
 * the License at
 * <p>
 * http://www.apache.org/licenses/LICENSE-2.0
 * <p>
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations under
 * the License.
 */

package io.renren.modules.sys.controller;


import io.renren.common.annotation.SysLog;
import io.renren.common.utils.PageUtils;
import io.renren.common.utils.R;
import io.renren.common.validator.Assert;
import io.renren.common.validator.ValidatorUtils;
import io.renren.common.validator.group.AddGroup;
import io.renren.common.validator.group.UpdateGroup;
import io.renren.common.validator.group.UpdateInfoGroyp;
import io.renren.common.validator.group.UserInfoAddGroup;
import io.renren.modules.sys.entity.SysUserEntity;
import io.renren.modules.sys.service.SysUserRoleService;
import io.renren.modules.sys.service.SysUserService;
import io.renren.modules.sys.shiro.ShiroUtils;
import io.renren.modules.sys.vo.SysUserVO;
import org.apache.commons.lang.ArrayUtils;
import org.apache.shiro.authz.annotation.RequiresPermissions;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import java.util.Map;
import java.util.regex.Matcher;
import java.util.regex.Pattern;

/**
 * 系统用户
 * 
 * @author chenshun
 * @email sunlightcs@gmail.com
 * @date 2016年10月31日 上午10:40:10
 */
@RestController
@RequestMapping("/sys/user")
public class SysUserController extends AbstractController {
	@Autowired
	private SysUserService sysUserService;
	@Autowired
	private SysUserRoleService sysUserRoleService;
	
	/**
	 * 所有用户列表
	 */
	@RequestMapping("/list")
	@RequiresPermissions("sys:user:list")
	public R list(@RequestParam Map<String, Object> params){
		PageUtils page = sysUserService.queryPage(params);

		return R.ok().put("page", page);
	}

	/**
	 * 用户申请列表
	 */
	@GetMapping("/applyList")
	public R applyList(@RequestParam Map<String, Object> params){
		PageUtils page = sysUserService.applyList(params);
        generatorUserVO(page);
        return R.ok().put("page", page);
	}
	/**
	 * 用户管理列表
	 */
	@GetMapping("/manageList")
	public R manageList(@RequestParam Map<String, Object> params){
		PageUtils page = sysUserService.manageList(params);
        generatorUserVO(page);
        return R.ok().put("page", page);
	}

    private void generatorUserVO(PageUtils page) {
        List<SysUserEntity> userEntityList = (List<SysUserEntity>) page.getList();
        List<SysUserVO> userVOList = new ArrayList<>();
        for (SysUserEntity userEntity : userEntityList) {
            SysUserVO userVO = new SysUserVO();
            userVO.setUserId(userEntity.getUserId());
            userVO.setName(userEntity.getName());
            userVO.setUsername(userEntity.getUsername());
            userVO.setJobNumber(userEntity.getJobNumber());
            userVO.setAudit(userEntity.getAudit());
            userVOList.add(userVO);
        }
        page.setList(userVOList);
    }

    /**
	 * 申请信息详情
	 */
	@GetMapping("/applyInfo/{userId}")
	public R applyInfo(@PathVariable("userId") Long userId){
	    SysUserVO userVO = sysUserService.applyInfo(userId);
		return R.ok().put("applyInfo",userVO);
	}

	/**
	 * 审核通过
	 */
	@GetMapping("/applyPass/{userId}")
	public R applyPass(@PathVariable("userId") Long userId){
	    sysUserService.applyPass(userId);
		return R.ok();
	}

	/**
	 * 获取登录的用户信息
	 */
	@RequestMapping("/info")
	public R info(){
		return R.ok().put("user", getUser());
	}
	
	/**
	 * 修改登录用户密码
	 */
	@SysLog("修改密码")
	@RequestMapping("/password")
	public R password(String password, String newPassword){
		Assert.isBlank(newPassword, "新密码不为能空");

		//原密码
		password = ShiroUtils.sha256(password, getUser().getSalt());
		//新密码
		newPassword = ShiroUtils.sha256(newPassword, getUser().getSalt());
				
		//更新密码
		boolean flag = sysUserService.updatePassword(getUserId(), password, newPassword);
		if(!flag){
			return R.error("原密码不正确");
		}
		
		return R.ok();
	}
	
	/**
	 * 用户信息
	 */
	@RequestMapping("/info/{userId}")
	@RequiresPermissions("sys:user:info")
	public R info(@PathVariable("userId") Long userId){
		SysUserEntity user = sysUserService.selectById(userId);
		
		//获取用户所属的角色列表
		List<Long> roleIdList = sysUserRoleService.queryRoleIdList(userId);
		user.setRoleIdList(roleIdList);
		
		return R.ok().put("user", user);
	}

	/**
	 * 保存用户
	 */
	@SysLog("保存用户")
	@RequestMapping("/save")
	@RequiresPermissions("sys:user:save")
	public R save(@RequestBody SysUserEntity user){
		ValidatorUtils.validateEntity(user, AddGroup.class);

		sysUserService.save(user);

		return R.ok();
	}
	
	/**
	 * 修改用户
	 */
	@SysLog("修改用户")
	@RequestMapping("/update")
	@RequiresPermissions("sys:user:update")
	public R update(@RequestBody SysUserEntity user){
		ValidatorUtils.validateEntity(user, UpdateGroup.class);

		sysUserService.update(user);
		
		return R.ok();
	}
	/**
	 * 修改信息
	 */
	@RequestMapping("/updateInfo")
	public R updateInfo(@RequestBody SysUserEntity user){
		ValidatorUtils.validateEntity(user, UpdateInfoGroyp.class);

		return sysUserService.updateInfo(user);
	}

	/**
	 * 填写个人信息
	 */
	@RequestMapping("/addUserInfo")
	public R addUserInfo(@RequestBody SysUserEntity user){
		ValidatorUtils.validateEntity(user, UserInfoAddGroup.class);
        if(user.getUserId() == null || user.getUserId() == 0){
            return R.error("填写个人信息失败");
        }

		sysUserService.addUserInfo(user);

		return R.ok();
	}
	
	/**
	 * 删除用户
	 */
	@SysLog("删除用户")
	@RequestMapping("/delete")
	@RequiresPermissions("sys:user:delete")
	public R delete(@RequestBody Long[] userIds){
		if(ArrayUtils.contains(userIds, 1L)){
			return R.error("系统管理员不能删除");
		}
		
		if(ArrayUtils.contains(userIds, getUserId())){
			return R.error("当前用户不能删除");
		}

		sysUserService.deleteBatchIds(Arrays.asList(userIds));
		
		return R.ok();
	}
}