# renren-admin/src/main/resources/public/swagger/api.yaml
swagger: '2.0'
info:
  description: 建科院平台项目后台接口文档
  version: 1.0.0
  title: 建科院平台项目
basePath: /
schemes:
- http

#定义接口数据
paths:
...
  /sys/dict/listByType:
    get:
      tags:
      - 字典数据管理
      summary: 指定类型的字典数据列表
      produces:
      - application/json
      parameters:
      - name: type
        description: 字典类型
        in: query
        type: string
        required: true
      responses:
        '200':
          description: 返回字典数据列表
          schema:
            $ref: '#/definitions/DictionaryVO'
  /sys/dict/infoByCode/{code}:
    get:
      tags:
      - 字典数据管理
      summary: 字典详情
      produces:
      - application/json
      parameters:
      - name: code
        description: 字典码
        in: path
        type: string
        required: true
      responses:
        '200':
          description: 返回字典详情
          schema:
            $ref: '#/definitions/DictionaryVO'
...
  /sys/user/addUserInfo:
    post:
      tags:
      - 登录
      summary: 信息填写
      produces:
      - application/json
      parameters:
      - name: body
        description: 用户信息对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/UserInfo'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/applyList:
    get:
      tags:
      - 系统管理模块
      summary: 用户申请列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: name
        description: 姓名
        in: query
        type: string
      - name: username
        description: 手机号
        in: query
        type: string
      - name: identity_card
        description: 身份证号
        in: query
        type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/SysUserListVO'
  /sys/user/manageList:
    get:
      tags:
      - 系统管理模块
      summary: 用户管理列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: name
        description: 姓名
        in: query
        type: string
      - name: username
        description: 手机号
        in: query
        type: string
      - name: jobNumber
        description: 工号
        in: query
        type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/SysUserListVO'
  /sys/user/applyInfo/{userId}:
    get:
      tags:
      - 系统管理模块
      summary: 用户申请详情
      produces:
      - application/json
      parameters:
      - name: userId
        description: 用户ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/ApplyInfo'
  /sys/user/applyPass/{userId}:
    get:
      tags:
      - 系统管理模块
      summary: 审核通过
      produces:
      - application/json
      parameters:
      - name: userId
        description: 用户ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/updateInfo:
    post:
      tags:
      - 系统管理模块
      summary: 用户信息修改
      produces:
      - application/json
      parameters:
      - name: body
        description: 用户对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/UserInfo_Update'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  DictionaryVO:
    type: object
    properties:
      name:
        description: 字典名称
        type: string
      type:
        description: 字典类型
        type: string
      code:
        description: 字典码
        type: string
      value:
        description: 字典码描述值
        type: string
      orderNum:
        description: 排序字段
        type: integer
        format: int32
...
  UserInfo:
    type: object
    properties:
      userId:
        description: 用户id
        type: integer
        format: int64
      name:
        description: 姓名
        type: string
      company:
        description: 公司名称
        type: string
      jobNumber:
        description: 工号
        type: string
      identityCard:
        description: 身份证
        type: string
      email:
        description: 邮箱
        type: string
  SysUserListVO:
    type: object
    properties:
      userId:
        description: 用户id
        type: integer
        format: int64
      name:
        description: 姓名
        type: string
      username:
        description: 手机号
        type: string
      email:
        description: 邮箱
        type: string
      audit:
        description: 审核状态，0-待审核，1-审核通过
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
  ApplyInfo:
    type: object
    properties:
      userId:
        description: 用户id
        type: integer
        format: int64
      name:
        description: 姓名
        type: string
      company:
        description: 公司名称
        type: string
      jobNumber:
        description: 工号
        type: string
      identityCard:
        description: 身份证
        type: string
      email:
        description: 邮箱
        type: string
      audit:
        description: 审核状态，0-待审核，1-审核通过
        type: integer
        format: int32
  UserInfo_Update:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      name:
        description: 姓名
        type: string
      company:
        description: 公司
        type: string
      jobNumber:
        description: 工号
        type: string
      identityCard:
        description: 身份证
        type: string
      email:
        description: 邮箱
        type: string
...